+++ packages/core-js/internals/engine-is-deno.js
@@ -1,0 +1,2 @@
+/* global Deno -- Deno case */
+module.exports = typeof Deno == 'object' && Deno && typeof Deno.version == 'object';